Rework country films
