+++ src/frontend/app/(pages)/country/[slug]/page.tsx
@@ -1,5 +1,4 @@
 import { PageTitle } from '@/components/custom/page-title';
-import { CountryFilmsTable } from '@/components/tables/country-films-table';
 import { DescriptionList } from '@/components/custom/description-list';
 import { DescriptionItem } from '@/components/custom/description-item';
 import { Tabs, TabsList, TabsTrigger, TabsContent } from '@/components/ui/tabs';
@@ -10,6 +9,12 @@ import { PreviousTable } from '@/components/tables/previous-table';
 import { ChartWrapper } from '@/components/charts/chart-wrapper';
 import { Controls } from '@/components/controls';
 import { StackedBarChart } from '@/components/charts/stacked-bar';
+import { Pagination } from '@/components/custom/pagination';
+import { DataTable } from '@/components/vendor/data-table';
+import { columns } from '@/components/tables/films';
+
+import { fetchCountryFilms } from '@/lib/api/dataFetching';
+import { paginate } from '@/lib/helpers/pagination';
 
 import {
 	fetchCountry,
@@ -143,3 +148,31 @@ export default async function Page({
 		</div>
 	);
 }
+
+/**
+ * @description Country Films List component
+ * @param {String} slug - Country slug
+ * @returns {JSX.Element}
+ * @example
+ * <CountryFilmsTable slug={slug} />
+ */
+export const CountryFilmsTable = async ({
+	slug,
+	pageIndex,
+}: {
+	slug: string;
+	pageIndex: number;
+}): Promise<JSX.Element> => {
+	const pageLimit = 15;
+
+	const data = await fetchCountryFilms(slug, pageIndex, pageLimit);
+
+	const pageNumbers = paginate(data!.count, pageIndex, pageLimit);
+
+	return (
+		<>
+			{data && <DataTable columns={columns} data={data.results} />}
+			<Pagination pages={pageNumbers} pageIndex={pageIndex} />
+		</>
+	);
+};
